Remove content-blocker-lib-ios folder and move related code under Client
- Adjust scripts where needed
- Adjust gitignore and swiftlint
