+++ Client/ContentBlocker/ContentBlockerGenerator/.swiftpm/xcode/package.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
