+++ Client/ContentBlocker/ContentBlockerGenerator/ContentBlockerGenerator.xcodeproj/project.xcworkspace/contents.xcworkspacedata
