+++ Client/ContentBlocker/ContentBlockerGenerator/Sources/ContentBlockerGenerator/ContentBlockerFileManager.swift
@@ -29,7 +29,7 @@ struct DefaultContentBlockerFileManager: ContentBlockerFileManager {
         let execIsFromCorrectDir = fileManager.fileExists(atPath: fileManager.currentDirectoryPath + "/Package.swift")
         self.rootDirectory = execIsFromCorrectDir ? fileManager.currentDirectoryPath : fallbackRootDirectoryPath
         self.outputDirectory = URL(fileURLWithPath: "\(rootDirectory)/../Lists")
-        self.inputDirectory = "\(rootDirectory)/../../shavar-prod-lists"
+        self.inputDirectory = "\(rootDirectory)/../../../shavar-prod-lists"
 
         createDirectory()
     }
