+++ Client/ContentBlocker/ContentBlockerGenerator/Sources/ContentBlockerGenerator/Entity.swift
