+++ Client/ContentBlocker/ContentBlockerGenerator/Sources/ContentBlockerGenerator/FileCategory.swift
