+++ Client/ContentBlocker/TrackingProtectionStats.js
