+++ content_blocker_update.sh
@@ -4,7 +4,6 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-
 # Version 107.0 hash
 SHAVAR_COMMIT_HASH="1d71be25893b05ba85850078b1ccedc0824ca958"
 
@@ -15,4 +14,4 @@ npm run build
 # Clone shavar prod list
 rm -rf shavar-prod-lists && git clone https://github.com/mozilla-services/shavar-prod-lists.git && git -C shavar-prod-lists checkout $SHAVAR_COMMIT_HASH
 
-(cd content-blocker-lib-ios/ContentBlockerGenerator && swift run)
+(cd Client/ContentBlocker/ContentBlockerGenerator && swift run)
